applied_python/lesson_6 Создание telegram бота/homework/requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
requests==2.32.3
matplotlib==3.9.4